+++ packages/ui-kit/src/common/Dropdown.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import { onClickOutside } from '@vueuse/core'
 import { onMounted, ref } from 'vue'
-import { useRouter } from 'vue-router'
+import { onBeforeRouteUpdate } from 'vue-router'
 
 const props = withDefaults(
   defineProps<{
@@ -16,9 +16,8 @@ const props = withDefaults(
 
 const rootEl = ref<HTMLElement | null>(null)
 const isDropdownOpen = ref(props.isOpenedByDefault)
-const router = useRouter()
 
-router.afterEach(() => {
+onBeforeRouteUpdate(() => {
   closeDropdown()
 })