+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.0-rc.4] - 2023-05-07
 ### Added
 - `toolkit` - increase gas estimates
 - `root` - `toasts` package
@@ -49,7 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release
 - `root` - `toolkit` package
 
-[Unreleased]: https://github.com/dl-tokene/webkit/compare/0.0.0-rc.3...HEAD
+[Unreleased]: https://github.com/dl-tokene/webkit/compare/0.0.0-rc.4...HEAD
+[0.0.0-rc.4]: https://github.com/dl-tokene/webkit/compare/0.0.0-rc.3...0.0.0-rc.4
 [0.0.0-rc.3]: https://github.com/dl-tokene/webkit/compare/0.0.0-rc.2...0.0.0-rc.3
 [0.0.0-rc.2]: https://github.com/dl-tokene/webkit/compare/0.0.0-rc.1...0.0.0-rc.2
 [0.0.0-rc.1]: https://github.com/dl-tokene/webkit/compare/0.0.0-rc.0...0.0.0-rc.1